feat: set tx types gas_limit to u128

+++ crates/primitives/src/transaction/compat.rs
@@ -19,7 +19,7 @@ impl FillTxEnv for TransactionSigned {
         tx_env.caller = sender;
         match self.as_ref() {
             Transaction::Legacy(tx) => {
-                tx_env.gas_limit = tx.gas_limit;
+                tx_env.gas_limit = tx.gas_limit as u64;
                 tx_env.gas_price = U256::from(tx.gas_price);
                 tx_env.gas_priority_fee = None;
                 tx_env.transact_to = tx.to;
@@ -32,7 +32,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.max_fee_per_blob_gas.take();
             }
             Transaction::Eip2930(tx) => {
-                tx_env.gas_limit = tx.gas_limit;
+                tx_env.gas_limit = tx.gas_limit as u64;
                 tx_env.gas_price = U256::from(tx.gas_price);
                 tx_env.gas_priority_fee = None;
                 tx_env.transact_to = tx.to;
@@ -45,7 +45,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.max_fee_per_blob_gas.take();
             }
             Transaction::Eip1559(tx) => {
-                tx_env.gas_limit = tx.gas_limit;
+                tx_env.gas_limit = tx.gas_limit as u64;
                 tx_env.gas_price = U256::from(tx.max_fee_per_gas);
                 tx_env.gas_priority_fee = Some(U256::from(tx.max_priority_fee_per_gas));
                 tx_env.transact_to = tx.to;
@@ -58,7 +58,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.max_fee_per_blob_gas.take();
             }
             Transaction::Eip4844(tx) => {
-                tx_env.gas_limit = tx.gas_limit;
+                tx_env.gas_limit = tx.gas_limit as u64;
                 tx_env.gas_price = U256::from(tx.max_fee_per_gas);
                 tx_env.gas_priority_fee = Some(U256::from(tx.max_priority_fee_per_gas));
                 tx_env.transact_to = TxKind::Call(tx.to);
@@ -71,7 +71,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.max_fee_per_blob_gas = Some(U256::from(tx.max_fee_per_blob_gas));
             }
             Transaction::Eip7702(tx) => {
-                tx_env.gas_limit = tx.gas_limit;
+                tx_env.gas_limit = tx.gas_limit as u64;
                 tx_env.gas_price = U256::from(tx.max_fee_per_gas);
                 tx_env.gas_priority_fee = Some(U256::from(tx.max_priority_fee_per_gas));
                 tx_env.transact_to = tx.to;
